8302369: Reduce the stack size of the C1 compiler
