+++ src/hotspot/share/c1/c1_CodeStubs.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1999, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1999, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -110,6 +110,8 @@ class CounterOverflowStub: public CodeStub {
 
 public:
   CounterOverflowStub(CodeEmitInfo* info, int bci, LIR_Opr method) :  _info(info), _bci(bci), _method(method) {
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
   }
 
   virtual void emit_code(LIR_Assembler* e);
@@ -166,9 +168,21 @@ class RangeCheckStub: public CodeStub {
 
  public:
   // For ArrayIndexOutOfBoundsException.
-  RangeCheckStub(CodeEmitInfo* info, LIR_Opr index, LIR_Opr array);
+  RangeCheckStub(CodeEmitInfo* info, LIR_Opr index, LIR_Opr array)
+    : _index(index), _array(array), _throw_index_out_of_bounds_exception(false) {
+    assert(info != NULL, "must have info");
+    _info = new CodeEmitInfo(info);
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
+  }
   // For IndexOutOfBoundsException.
-  RangeCheckStub(CodeEmitInfo* info, LIR_Opr index);
+  RangeCheckStub(CodeEmitInfo* info, LIR_Opr index)
+    : _index(index), _array(), _throw_index_out_of_bounds_exception(true) {
+    assert(info != NULL, "must have info");
+    _info = new CodeEmitInfo(info);
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
+  }
   virtual void emit_code(LIR_Assembler* e);
   virtual CodeEmitInfo* info() const             { return _info; }
   virtual bool is_exception_throw_stub() const   { return true; }
@@ -335,7 +349,12 @@ class MonitorEnterStub: public MonitorAccessStub {
   CodeEmitInfo* _info;
 
  public:
-  MonitorEnterStub(LIR_Opr obj_reg, LIR_Opr lock_reg, CodeEmitInfo* info);
+  MonitorEnterStub(LIR_Opr obj_reg, LIR_Opr lock_reg, CodeEmitInfo* info)
+    : MonitorAccessStub(obj_reg, lock_reg) {
+    _info = new CodeEmitInfo(info);
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
+  }
 
   virtual void emit_code(LIR_Assembler* e);
   virtual CodeEmitInfo* info() const             { return _info; }
@@ -476,7 +495,10 @@ private:
 
 public:
   DeoptimizeStub(CodeEmitInfo* info, Deoptimization::DeoptReason reason, Deoptimization::DeoptAction action) :
-    _info(new CodeEmitInfo(info)), _trap_request(Deoptimization::make_trap_request(reason, action)) {}
+    _info(new CodeEmitInfo(info)), _trap_request(Deoptimization::make_trap_request(reason, action)) {
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
+  }
 
   virtual void emit_code(LIR_Assembler* e);
   virtual CodeEmitInfo* info() const           { return _info; }
@@ -499,6 +521,8 @@ class SimpleExceptionStub: public CodeStub {
  public:
   SimpleExceptionStub(Runtime1::StubID stub, LIR_Opr obj, CodeEmitInfo* info):
     _obj(obj), _stub(stub), _info(info) {
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(2 * BytesPerWord);
   }
 
   void set_obj(LIR_Opr obj) {
@@ -534,7 +558,10 @@ class ArrayCopyStub: public CodeStub {
   LIR_OpArrayCopy* _op;
 
  public:
-  ArrayCopyStub(LIR_OpArrayCopy* op): _op(op) { }
+  ArrayCopyStub(LIR_OpArrayCopy* op): _op(op) {
+    FrameMap* f = Compilation::current()->frame_map();
+    f->update_reserved_argument_area_size(arraycopystub_reserved_argument_area_size * BytesPerWord);
+  }
 
   LIR_Opr src() const                         { return _op->src(); }
   LIR_Opr src_pos() const                     { return _op->src_pos(); }
