+++ src/hotspot/share/c1/c1_Defs.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2000, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -59,6 +59,11 @@ enum {
 // true if LIR requires src1 and dst to match in binary LIR ops
 enum {
   two_operand_lir_form = pd_two_operand_lir_form
+};
+
+// the number of stack required by ArrayCopyStub
+enum {
+  arraycopystub_reserved_argument_area_size = pd_arraycopystub_reserved_argument_area_size
 };
 
 #endif // SHARE_C1_C1_DEFS_HPP
